move to test_raftstore
