+++ kv/test_raftstore/mock_cluster.go
@@ -1,4 +1,4 @@
-package inner_server
+package test_raftstore
 
 import (
 	"context"
@@ -14,44 +14,10 @@ import (
 	"github.com/pingcap-incubator/tinykv/kv/pd"
 	tikvConf "github.com/pingcap-incubator/tinykv/kv/tikv/config"
 	"github.com/pingcap-incubator/tinykv/kv/tikv/raftstore"
-	"github.com/pingcap-incubator/tinykv/kv/tikv/raftstore/message"
 	"github.com/pingcap-incubator/tinykv/kv/tikv/raftstore/snap"
 	"github.com/pingcap-incubator/tinykv/kv/tikv/worker"
 	"github.com/pingcap-incubator/tinykv/proto/pkg/metapb"
-	"github.com/pingcap-incubator/tinykv/proto/pkg/raft_serverpb"
 )
-
-type MockTransport struct {
-	sync.Mutex
-	routers map[uint64]message.RaftRouter
-}
-
-func NewMockTransport() MockTransport {
-	return MockTransport{
-		routers: make(map[uint64]message.RaftRouter),
-	}
-}
-
-func (t *MockTransport) AddNode(storeID uint64, raftRouter message.RaftRouter) {
-	t.Lock()
-	defer t.Unlock()
-
-	t.routers[storeID] = raftRouter
-}
-
-func (t *MockTransport) Send(msg *raft_serverpb.RaftMessage) error {
-	t.Lock()
-	defer t.Unlock()
-
-	storeID := msg.GetToPeer().GetStoreId()
-	router, ok := t.routers[storeID]
-	if !ok {
-		panic("storeID not found")
-	}
-	router.SendRaftMessage(msg)
-
-	return nil
-}
 
 type MockCluster struct {
 	trans    *MockTransport
